requirements.txt
# Core LLM + embeddings
transformers>=4.41.0
sentence-transformers>=3.0.0
accelerate>=0.31.0
safetensors>=0.4.3

# Quantization / memory helpers (very useful for 7B on consumer GPUs)
bitsandbytes>=0.43.0

# Vector store / retrieval
faiss-cpu>=1.7.4
numpy>=1.26.0
scipy>=1.11.0

# PDF ingestion
pypdf2>=3.0.1

# Backend API (if you're using FastAPI)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# Simple GUI
gradio>=4.44.0

# Utilities
python-dotenv>=1.0.1
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2
nltk>=3.8.1
datasets>=2.19.0

# Testing
pytest>=8.2.0
